+++ server/src/table_notify.go
@@ -73,7 +73,7 @@ func (t *Table) NotifyPlayerChange() {
 
 		// Second, send information about the game and the players in one big message
 		type GameMessage struct {
-			TableID           int                  `json:"tableID"`
+			TableID           uint64               `json:"tableID"`
 			Name              string               `json:"name"`
 			Owner             int                  `json:"owner"`
 			Players           []*GamePlayerMessage `json:"players"`
@@ -207,7 +207,7 @@ func (t *Table) NotifySound() {
 
 func (t *Table) NotifyFinishOngoingGame() {
 	type FinishOngoingGameMessage struct {
-		TableID            int    `json:"tableID"`
+		TableID            uint64 `json:"tableID"`
 		DatabaseID         int    `json:"databaseID"`
 		SharedReplayLeader string `json:"sharedReplayLeader"`
 	}
@@ -290,7 +290,7 @@ func (t *Table) NotifySpectatorsNote(order int) {
 		}
 
 		type NoteMessage struct {
-			TableID int `json:"tableID"`
+			TableID uint64 `json:"tableID"`
 			// The order of the card in the deck that these notes correspond to
 			Order int    `json:"order"`
 			Notes []Note `json:"notes"`
